Rename dashboardAux to dashboardContexts
Also tighten dep-cruiser rules

JIRA: RAIL-3383
