+++ libs/sdk-ui-dashboard/src/layout/DashboardLayout.tsx
@@ -1,6 +1,6 @@
 // (C) 2020 GoodData Corporation
 import React from "react";
-import { useDashboardComponentsContext } from "../dashboardAux";
+import { useDashboardComponentsContext } from "../dashboardContexts";
 import { DashboardLayoutProps } from "./types";
 
 /**
