+++ libs/sdk-ui-dashboard/src/layout/DefaultDashboardLayout.tsx
@@ -20,7 +20,7 @@ import {
 import { EmptyDashboardError } from "./EmptyDashboardError";
 import { selectInsights, selectSettings, selectBasicLayout, useDashboardSelector } from "../model";
 import { DashboardLayoutWidget } from "./DashboardLayoutWidget";
-import { useDashboardComponentsContext } from "../dashboardAux";
+import { useDashboardComponentsContext } from "../dashboardContexts";
 import { DashboardLayoutProps } from "./types";
 import {
     DashboardLayout,
