+++ libs/sdk-ui-dashboard/src/widget/insight/DashboardInsight.tsx
@@ -1,7 +1,7 @@
 // (C) 2020 GoodData Corporation
 import React from "react";
 
-import { useDashboardComponentsContext } from "../../dashboardAux";
+import { useDashboardComponentsContext } from "../../dashboardContexts";
 
 import { IDashboardInsightCoreProps } from "./types";
