+++ libs/sdk-ui-dashboard/src/widget/kpi/DefaultDashboardKpi/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useBackendStrict, useWorkspaceStrict } from "@gooddata/sdk-ui";
 import invariant from "ts-invariant";
 
-import { useDashboardComponentsContext } from "../../../dashboardAux";
+import { useDashboardComponentsContext } from "../../../dashboardContexts";
 import {
     useDashboardSelector,
     selectSeparators,
